Milestone: End-to-end testing with Mandarin audio transcription
- Fixed OAuth2 token request in test script
- Improved category enum handling in audio service
- Added eager loading for chat history relationships
- Enhanced error logging throughout the system
- Successfully tested full pipeline:
  * Mandarin audio transcription
  * Text categorization
  * Database storage
  * Retrieval of categorized entries

+++ app/schemas.py
@@ -27,7 +27,7 @@ class TaskResponse(TaskBase):
     user_id: int
     start_time: datetime
     end_time: Optional[datetime] = None
-    duration: Optional[float] = None
+    duration: Optional[int] = None  # Duration in seconds
     model_config = ConfigDict(from_attributes=True)
 
 class TaskSummary(BaseModel):
